OTF code cleanup
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1786659 13f79535-47bb-0310-9956-ffa450edef68
